(fix) Services Logic

+++ backend/app/services/user.py
@@ -1,6 +1,12 @@
 from sqlalchemy import select
 from sqlalchemy.orm import Session
+
+from backend.app.models import Booking
 from backend.app.models.user import User
+from fastapi import HTTPException
+
+from backend.app.security.auth import hash_password, verify_password
+
 
 class UserService:
     def __init__(self, session: Session):
@@ -8,7 +14,7 @@ class UserService:
 
     def add_user(self, name: str, email: str, password: str, role: str):
         if self.find_user_by_email(email):
-            raise ValueError("User with this email already exists")
+            raise HTTPException(status_code=400, detail="Email already registered")
         user = User(name=name, email=email, password=password, role=role)
         self.session.add(user)
         self.session.commit()
@@ -18,43 +24,67 @@ class UserService:
     def delete_user(self, user_id: int) -> bool:
         user = self.get_user_by_id(user_id)
         if user:
+            active_bookings = self.session.scalars(
+                select(Booking)
+                .where(Booking.user_id == user_id, Booking.status.in_(["confirmed", "pending"]))
+            )
+            if active_bookings:
+                raise HTTPException(status_code=400, detail="User has active bookings!")
             self.session.delete(user)
             self.session.commit()
             return True
         return False
 
-    def find_user_by_email(self, email: str):
+    def find_user_by_email(self, email: str) -> User | None:
         return self.session.scalars(
             select(User).where(User.email == email)
         ).one_or_none()
 
     def get_user_by_email(self, email: str) -> User:
-        email_user = self.find_user_by_email(email)
-        if not email_user:
-            raise ValueError("User not found")
-        return email_user
+        user = self.find_user_by_email(email)
+        if not user:
+            raise HTTPException(status_code=404, detail="User not found")
+        return user
 
     def exists_user_email(self, email: str) -> bool:
         return self.find_user_by_email(email) is not None
 
     def get_user_by_id(self, user_id: int) -> User:
-        id_user = self.session.scalars(select(User).where(User.id == user_id)).first()
-        return id_user
+        user = self.session.scalars(
+            select(User)
+            .where(User.id == user_id)).first()
+        return user
 
     def get_users_by_name(self, name: str) -> list[User]:
-        name_user = self.session.scalars(select(User).where(User.name == name)).all()
-        return list(name_user)
+        users = self.session.scalars(select(User).where(User.name.ilike(f"%{name.strip()}%"))).all()
+        if not users:
+            raise HTTPException(status_code=404, detail="User not found")
+        return list(users)
 
     def get_users(self) -> list[User]:
         all_users = self.session.scalars(select(User)).all()
         return list(all_users)
 
     def register_user(self, name: str, email: str, password: str, role = "USER"):
-        return self.add_user(name, email, password ,role)
+        new_password = hash_password(password)
+        return self.add_user(name, email, new_password ,role)
+
+    def login_user(self, email: str, password: str) -> User:
+        user = self.get_user_by_email(email)
+        if not user:
+            raise HTTPException(status_code=404, detail="User not found")
+        if self.check_password(user.id, password):
+            return user
+        raise HTTPException(
+            status_code=401,
+            detail="Incorrect email or password",
+        )
 
     def check_password(self,user_id: int, password: str) -> bool:
         user = self.get_user_by_id(user_id)
-        return user.password == password
+        if not user:
+            raise HTTPException(status_code=404, detail="User not found")
+        return verify_password(password, user.password)
 
     def edit_user(self, edit: dict) -> User:
         if "id" not in edit:
